Organize action and api functions in different files
